perf spike: E1-E4 hot-path experiments (-37%..-48% LALR parse wall-clock)
Four measured experiments on the LALR hot path (each suite-green, 675/675):

- E1 token plumbing: TokenSource::peek_type/take_current — dispatch reads the
  cached token's id, SHIFT moves the token out; kills the per-token peek-clone
  and SHIFT-clone (each token was materialized 3x = 6 String allocs, now 1x).
- E2 dense per-token tables: terminal-name Vec (was HashMap<SymbolId,String>),
  contextual state->scanner Vec (was HashMap<usize,usize>), %ignore bitset
  (was HashSet) — SipHash is gone from the parse profile (5% -> 0.5%).
- E3 in-place reduction: shape_reduction drains the value stack tail itself;
  no per-reduction drain().collect() Vec.
- E4 u32 positions: Token position fields + Meta fields usize -> u32; the
  generic stack element (GSlot<Child>) shrinks 264 -> ~168 bytes. Public field
  type change - a measurement for the ARCH "default Tree gets cheaper"
  decision (performance-strategy.md §5), not a unilateral landing.

json_large 17.67 -> 11.10 ms (5.2 -> 8.3 MB/s), arith_large 803 -> 416 us,
NullBuilder floor share of parse() 59.5% -> 46.9%. Details in
docs/notes/perf-spike-2026-07-01.md.

+++ lark-rs/src/error.rs
@@ -167,16 +167,16 @@ impl ParseError {
     ) -> ParseError {
         if token.type_id == crate::grammar::intern::SymbolId::END {
             ParseError::UnexpectedEof {
-                line: token.line,
-                col: token.column,
+                line: token.line as usize,
+                col: token.column as usize,
                 expected,
             }
         } else {
             ParseError::UnexpectedToken {
                 token: token.value.clone(),
                 token_type: token.type_.clone(),
-                line: token.line,
-                col: token.column,
+                line: token.line as usize,
+                col: token.column as usize,
                 expected,
             }
         }
@@ -199,8 +199,8 @@ impl ParseError {
         ParseError::UnexpectedToken {
             token: token.value.clone(),
             token_type: token.type_.clone(),
-            line: token.line,
-            col: token.column,
+            line: token.line as usize,
+            col: token.column as usize,
             expected,
         }
     }
